Keep only one CLI
