--- src/cloudai/__main__.py
@@ -1,29 +1,0 @@
-# SPDX-FileCopyrightText: NVIDIA CORPORATION & AFFILIATES
-# Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from cloudai.cli import CloudAICLI, setup_logging
-
-
-def main() -> None:
-    cloudai_cli = CloudAICLI()
-    args = cloudai_cli.init_default_args().parse_args()
-    setup_logging(args.log_file, args.log_level)
-    rc = cloudai_cli.run()
-    exit(rc)
-
-
-if __name__ == "__main__":
-    main()
